Make CloseIdleHandler sharable
