+++ vertx-core/src/main/java/org/vertx/java/core/http/impl/DefaultHttpServer.java
@@ -30,8 +30,6 @@ import io.netty.handler.codec.http.websocketx.WebSocketServerHandshaker;
 import io.netty.handler.codec.http.websocketx.WebSocketServerHandshakerFactory;
 import io.netty.handler.ssl.SslHandler;
 import io.netty.handler.stream.ChunkedWriteHandler;
-import io.netty.handler.timeout.IdleState;
-import io.netty.handler.timeout.IdleStateEvent;
 import io.netty.handler.timeout.IdleStateHandler;
 import io.netty.util.CharsetUtil;
 import io.netty.util.concurrent.GlobalEventExecutor;
@@ -196,17 +194,7 @@ public class DefaultHttpServer implements HttpServer, Closeable {
                 // Only allIdleTime is enabled (different from 0) to trigger an event
                 // when neither read nor write was performed for the specified period of time.
                 pipeline.addLast("idle", new IdleStateHandler(0, 0, tcpHelper.getIdleTimeout()));
-                pipeline.addLast("closeIdle", new ChannelDuplexHandler() {
-                  @Override
-                  public void userEventTriggered(ChannelHandlerContext ctx, Object evt) throws Exception {
-                    if (evt instanceof IdleStateEvent) {
-                      IdleStateEvent e = (IdleStateEvent) evt;
-                      if (e.state() == IdleState.ALL_IDLE) {
-                        ctx.close();
-                      }
-                    }
-                  }
-                });
+                pipeline.addLast("closeIdle", CloseIdleHandler.INSTANCE);
               }
 
               pipeline.addLast("flashpolicy", new FlashPolicyHandler());
