+++ vertx-core/src/main/java/org/vertx/java/core/net/impl/CloseIdleHandler.java
@@ -1,0 +1,48 @@
+/*
+ * Copyright (c) 2011-2013 The original author or authors
+ * ------------------------------------------------------
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v1.0
+ * and Apache License v2.0 which accompanies this distribution.
+ *
+ *     The Eclipse Public License is available at
+ *     http://www.eclipse.org/legal/epl-v10.html
+ *
+ *     The Apache License v2.0 is available at
+ *     http://www.opensource.org/licenses/apache2.0.php
+ *
+ * You may elect to redistribute this code under either of these licenses.
+ */
+
+package org.vertx.java.core.net.impl;
+
+import io.netty.channel.ChannelDuplexHandler;
+import io.netty.channel.ChannelHandlerContext;
+import io.netty.channel.ChannelHandler.Sharable;
+import io.netty.handler.timeout.IdleState;
+import io.netty.handler.timeout.IdleStateEvent;
+import io.netty.handler.timeout.IdleStateHandler;
+
+/**
+ * Netty handler to receive events for idle connections generated by innactivity in a TCP
+ * connection detected by {@link IdleStateHandler}. The handler closes the idle connection.
+ */
+@Sharable
+public class CloseIdleHandler extends ChannelDuplexHandler {
+
+  /**
+   * Static instance to be shared by multiple netty pipelines.
+   */
+  public static CloseIdleHandler INSTANCE = new CloseIdleHandler();
+
+  @Override
+  public void userEventTriggered(ChannelHandlerContext ctx, Object evt) throws Exception {
+    if (evt instanceof IdleStateEvent) {
+      IdleStateEvent e = (IdleStateEvent) evt;
+      if (e.state() == IdleState.ALL_IDLE) {
+        ctx.close();
+      }
+    }
+  }
+
+}
