+++ vertx-core/src/main/java/org/vertx/java/core/net/impl/DefaultNetServer.java
@@ -26,8 +26,6 @@ import io.netty.channel.group.DefaultChannelGroup;
 import io.netty.channel.socket.nio.NioServerSocketChannel;
 import io.netty.handler.ssl.SslHandler;
 import io.netty.handler.stream.ChunkedWriteHandler;
-import io.netty.handler.timeout.IdleState;
-import io.netty.handler.timeout.IdleStateEvent;
 import io.netty.handler.timeout.IdleStateHandler;
 import io.netty.util.concurrent.Future;
 import io.netty.util.concurrent.GenericFutureListener;
@@ -138,17 +136,7 @@ public class DefaultNetServer implements NetServer, Closeable {
               // Only allIdleTime is enabled (different from 0) to trigger an event
               // when neither read nor write was performed for the specified period of time.
               pipeline.addLast("idle", new IdleStateHandler(0, 0, tcpHelper.getIdleTimeout()));
-              pipeline.addLast("closeIdle", new ChannelDuplexHandler() {
-                @Override
-                public void userEventTriggered(ChannelHandlerContext ctx, Object evt) throws Exception {
-                  if (evt instanceof IdleStateEvent) {
-                    IdleStateEvent e = (IdleStateEvent) evt;
-                    if (e.state() == IdleState.ALL_IDLE) {
-                      ctx.close();
-                    }
-                  }
-                }
-              });
+              pipeline.addLast("closeIdle", CloseIdleHandler.INSTANCE);
             }
 
             if (tcpHelper.isSSL()) {
